refactor: refactor authentication package
