+++ packages/authentication/maker.go
@@ -1,0 +1,6 @@
+package authentication
+
+type Maker interface {
+	Generate(user *User) (string, error)
+	Verify(token string) (*User, error)
+}
